Prevent double click on saving bread
Double clicking save button causes record to be saved two times.

This PR disables button on submit preventing form to be submitted two times.
Button does not need to be re-enabled on error since validation is not done with ajax.
In case of custom button it should contain the _save_ class to make use of this functionality.